# Improve performance of stream compare