--- a/src/StyleCopTests.ruleset
+++ b/src/StyleCopTests.ruleset
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RuleSet Name="Rules for Yarhl.IntegrationTests" ToolsVersion="12.0">
-  <Include Path="..\StyleCop.ruleset" Action="Default" />
+<RuleSet Name="Rules for Yarhl tests" ToolsVersion="12.0">
+  <Include Path="StyleCop.ruleset" Action="Default" />
   <Rules AnalyzerId="Microsoft.Analyzers.ManagedCodeAnalysis" RuleNamespace="Microsoft.Rules.Managed">
     <!-- Disable documentation in unit tests. -->
     <Rule Id="SA0001" Action="None" />
+    <Rule Id="SA1600" Action="None" />
   </Rules>
 </RuleSet>