--- a/src/Yarhl.PerformanceTests/Program.cs
+++ b/src/Yarhl.PerformanceTests/Program.cs
@@ -1,0 +1,55 @@
+﻿// Copyright (c) 2019 Benito Palacios Sánchez
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+namespace Yarhl.PerformanceTests
+{
+    using BenchmarkDotNet.Running;
+
+    static class Program
+    {
+        static void Main(string[] args)
+        {
+            if (args.Length > 0 && args[0] == "auto") {
+                RunAuto();
+            } else {
+                RunManual();
+            }
+        }
+
+        static void RunAuto()
+        {
+            BenchmarkRunner.Run<DataStreamCompare>();
+        }
+
+        static void RunManual()
+        {
+            const int Iterations = 1_000;
+
+            var test1 = new DataStreamCompare();
+            test1.Length = 1024 * 1024;
+            test1.SetUp();
+
+            for (int i = 0; i < Iterations; i++) {
+                test1.Compare();
+            }
+
+            test1.CleanUp();
+        }
+    }
+}