--- a/src/Yarhl.UnitTests/StyleCop.ruleset
+++ b/src/Yarhl.UnitTests/StyleCop.ruleset
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RuleSet Name="Rules for Yarhl.UnitTests" ToolsVersion="12.0">
-  <Include Path="..\StyleCop.ruleset" Action="Default" />
-  <Rules AnalyzerId="Microsoft.Analyzers.ManagedCodeAnalysis" RuleNamespace="Microsoft.Rules.Managed">
-    <!-- Disable documentation in unit tests. -->
-    <Rule Id="SA0001" Action="None" />
-    <Rule Id="SA1600" Action="None" />
-  </Rules>
-</RuleSet>