--- a/src/Yarhl.sln
+++ b/src/Yarhl.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yarhl.UnitTests", "Yarhl.Un
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yarhl.IntegrationTests", "Yarhl.IntegrationTests\Yarhl.IntegrationTests.csproj", "{B32EE98E-01F2-4B98-8160-7567C5AE1A8D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yarhl.PerformanceTests", "Yarhl.PerformanceTests\Yarhl.PerformanceTests.csproj", "{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -72,5 +74,17 @@ Global
 		{B32EE98E-01F2-4B98-8160-7567C5AE1A8D}.Release|x64.Build.0 = Release|Any CPU
 		{B32EE98E-01F2-4B98-8160-7567C5AE1A8D}.Release|x86.ActiveCfg = Release|Any CPU
 		{B32EE98E-01F2-4B98-8160-7567C5AE1A8D}.Release|x86.Build.0 = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|x64.Build.0 = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Debug|x86.Build.0 = Debug|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|x64.ActiveCfg = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|x64.Build.0 = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|x86.ActiveCfg = Release|Any CPU
+		{ADF44DD6-B355-45FB-A50A-06CBCE2EFF6E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal