--- a/src/Yarhl/IO/DataStream.cs
+++ b/src/Yarhl/IO/DataStream.cs
@@ -560,18 +560,30 @@ namespace Yarhl.IO
             if (otherStream.Disposed)
                 throw new ObjectDisposedException(nameof(otherStream));
 
-            if (Length != otherStream.Length)
+            if (Length != otherStream.Length) {
                 return false;
+            }
 
             long startPosition = Position;
             long otherStreamStartPosition = otherStream.position;
             Seek(0, SeekMode.Start);
             otherStream.Seek(0, SeekMode.Start);
 
+            const int BufferSize = 70 * 1024;
+            byte[] buffer1 = new byte[BufferSize];
+            byte[] buffer2 = new byte[BufferSize];
+
             bool result = true;
             while (!EndOfStream && result) {
-                if (ReadByte() != otherStream.ReadByte())
-                    result = false;
+                int length = (int)(Position + BufferSize > Length ? Length - Position : BufferSize);
+                Read(buffer1, 0, length);
+                otherStream.Read(buffer2, 0, length);
+
+                for (int i = 0; i < length && result; i++) {
+                    if (buffer1[i] != buffer2[i]) {
+                        result = false;
+                    }
+                }
             }
 
             Seek(startPosition, SeekMode.Start);