### Description
Comparing large streams was not feasible since the `Compare` method was taking too long. Instead of reading byte by byte now we read by blocks. I have added also a benchmark project so we can start writing more benchmark and compare between releases.

For the use case of 10 MB, before it was taking `16,357 us` and now it takes `850 us`.

As part of the pull request I have added also a configuration file with code style settings that VS can use.

### Performan results

#### Before
|  Method |  Length |            Mean |         Error |        StdDev |
|-------- |-------- |----------------:|--------------:|--------------:|
| Compare |      32 |        523.5 ns |      1.325 ns |      1.174 ns |
| Compare |    1024 |     16,013.9 ns |     26.025 ns |     23.071 ns |
| Compare |    5120 |     81,648.6 ns |    106.389 ns |     94.311 ns |
| Compare |   81920 |  1,372,969.3 ns |  2,081.951 ns |  1,625.450 ns |
| Compare | 1048576 | 16,357,543.1 ns | 23,864.284 ns | 22,322.666 ns |

#### Trying to replace look with `Enumerable.SequenceEqual`

|  Method |  Length |       Mean |      Error |     StdDev |
|-------- |-------- |-----------:|-----------:|-----------:|
| Compare |      32 |   626.2 us |  1.5622 us |  1.4613 us |
| Compare |    1024 |   624.9 us |  0.4670 us |  0.4369 us |
| Compare |    5120 |   626.9 us |  2.9577 us |  2.6219 us |
| Compare |   81920 | 1,256.7 us |  3.2089 us |  2.8446 us |
| Compare | 1048576 | 9,403.7 us | 21.5151 us | 19.0726 us |

#### Final implementation

|  Method |  Length |       Mean |     Error |    StdDev |
|-------- |-------- |-----------:|----------:|----------:|
| Compare |      32 |   4.936 us | 0.0239 us | 0.0200 us |
| Compare |    1024 |   5.990 us | 0.0224 us | 0.0210 us |
| Compare |    5120 |   9.170 us | 0.0163 us | 0.0136 us |
| Compare |   81920 |  70.103 us | 0.1514 us | 0.1342 us |
| Compare | 1048576 | 850.875 us | 0.8558 us | 0.8005 us |

Now we can event tests large streams of 512 MB (was taking too long before to write a test).

|  Method |    Length |           Mean |         Error |        StdDev |         Median |
|-------- |---------- |---------------:|--------------:|--------------:|---------------:|
| Compare | 536870912 | 454,487.979 us | 1,555.8960 us | 1,379.2613 us | 454,634.450 us |

**Note**: This tests are comparing in-memory streams which is way faster than comparing a stream from the disk. In that case, the performance hit is on the read operation over the disk.